test(sessions): split target dedupe regressions
